GOST Checker/requirements.txt
python-docx>=0.8.11
pyyaml>=5.4.1